Moved caloric intake tracker to data-tracker
