+++ src/app/data-tracker/activity-tracker/activity-tracker.component.html
@@ -1,17 +1,18 @@
 <div>
   <h2>Activity Tracker</h2>
-  <label for="exerciseName">Exercise Name:</label>
-  <input id="exerciseName" [(ngModel)]="exercise.name" />
+  <ul>
+  <li><label for="exerciseName">Exercise Name:</label>
+  <input id="exerciseName" [(ngModel)]="exercise.name" /></li>
 
-  <label for="exerciseNotes">Notes:</label>
-  <textarea id="exerciseNotes" [(ngModel)]="exercise.notes"></textarea>
+  <li><label for="exerciseNotes">Notes:</label>
+  <textarea id="exerciseNotes" [(ngModel)]="exercise.notes"></textarea></li>
 
-  <label for="caloriesBurned">Calories Burned:</label>
+  <li><label for="caloriesBurned">Calories Burned:</label>
   <input
     type="number"
     id="caloriesBurned"
-    [(ngModel)]="exercise.caloriesBurned"
-  />
+    [(ngModel)]="exercise.caloriesBurned"/></li>
 
-  <button (click)="addActivity()">Add Activity</button>
+  <li><button (click)="addActivity()">Add Activity</button></li>
+</ul>
 </div>
